requirements.txt

streamlit==1.39.0
google-generativeai==0.8.3
reportlab==4.2.5
python-dotenv==1.0.1
